Renaming JenisBeasiswa to BeasiswaTypes and moving truncateText function to /lib/utils.ts
