+++ src/app/beasiswa/page.tsx
@@ -1,14 +1,14 @@
-import BeasiswaHeader from '@/components/Beasiswa/Header/Header'
-import JenisBeasiswa from '@/components/Beasiswa/JenisBeasiswa/JenisBeasiswa'
-import React from 'react'
+import BeasiswaHeader from "@/components/Beasiswa/Header/Header";
+import JenisBeasiswa from "@/components/Beasiswa/BeasiswaTypes/BeasiswaTypes";
+import React from "react";
 
 function page() {
   return (
     <div>
-        <BeasiswaHeader />
-        <JenisBeasiswa/>
+      <BeasiswaHeader />
+      <JenisBeasiswa />
     </div>
-  )
+  );
 }
 
-export default page
+export default page;
