+++ src/components/Beasiswa/BeasiswaTypes/BeasiswaTypes.tsx
